requirements/requirements.txt
click==8.1.7
opencv-python==4.13.0.92
Pillow==12.1.1
numpy==2.4.2
scipy==1.17.1
tqdm==4.67.3
tabulate==0.10.0
scikit-learn==1.8.0
joblib==1.5.3
matplotlib==3.10.8
seaborn==0.13.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5-sip==12.18.0
reportlab==4.4.10


pywin32==311; sys_platform == 'win32'
